+++ CHANGELOG.md
@@ -17,6 +17,8 @@ CHANGELOG for FlatCAM beta
 - Fixed a recent change that made the edited Geometry objects in the Geometry Editor not to be plotted after saving changes
 - modified the Tool Database such that when a tool shape is selected as 'V' any change in the Vdia or Vangle or CutZ parameters will update the tool diameter value
 - In Tool Isolation made sure that the use of ESC key while some processes are active will disconnect the mouse events that may be connected, correctly
+- optimized the Gerber UI
+- added a Multi-color checkbox for the Geometry UI (will color differently tool geometry when the geometry is multitool)
 
 29.05.2020